requirements.txt
# Motorsport Digest - Python Dependencies
# Version: 1.0.0
# Python: 3.9+

# ============================================
# CORE DEPENDENCIES
# ============================================

# RSS Feed Parsing
feedparser==6.0.11

# Article Extraction & Web Scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# AI Summarization
anthropic==0.18.1

# Alternative AI (optional - uncomment if using OpenAI)
# openai==1.12.0

# ============================================
# UTILITIES
# ============================================

# Date & Time Handling
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.1

# Progress Bars (optional)
tqdm==4.66.1

# ============================================
# WEB GENERATION
# ============================================

# Templating (if needed for advanced HTML)
jinja2==3.1.3

# Markdown Support (optional)
markdown==3.5.2

# ============================================
# DATABASE
# ============================================

# SQLite is built-in to Python, no need to install

# ============================================
# TESTING & DEVELOPMENT (optional)
# ============================================

# Testing
# pytest==8.0.0
# pytest-cov==4.1.0

# Code Quality
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0

# ============================================
# DEPLOYMENT & AUTOMATION
# ============================================

# Already handled by GitHub Actions
# No additional dependencies needed

# ============================================
# NOTES
# ============================================

# Install all dependencies:
#   pip install -r requirements.txt

# Install in virtual environment (recommended):
#   python -m venv venv
#   source venv/bin/activate  # Linux/Mac
#   venv\Scripts\activate     # Windows
#   pip install -r requirements.txt

# Freeze current environment:
#   pip freeze > requirements-frozen.txt

# Update dependencies:
#   pip install --upgrade -r requirements.txt
